+++ juriscraper/opinions/united_states/state/txctapp3.py
@@ -1,0 +1,17 @@
+# Scraper for Texas 3rd Court of Appeals
+# CourtID: txctapp3
+# Court Short Name: TX
+# Author: Andrei Chelaru
+# Reviewer: mlr
+# Date: 2014-07-10
+
+
+from juriscraper.opinions.united_states.state import texapp
+
+
+class Site(texapp.Site):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.court_id = self.__module__
+        self.court_number = "03"
+        self.checkbox = 4